feat: create a base functionality for the map enrichment node
The map enrichment node is responsible for dynamically insertion and removal of sign footprints on a static map.

Sign footprint is a virtual wall displayed only on the static map. It represents a line that goes through the top edge of the sign and expands (in both directions from the sign origin) until the nearest occupied cell (in each direction) of the static map.