feat: connect the map enrichment node to navigation stack
Currently the map enrichment node plays a role of a proxy between map_server and move_base. Therefore, a navigation stack has 2 maps. The first one is a original map of map_server. It is needed for the amcl node that uses the original map of environment for AGV localization. It is assumed the real sign and its virtual footprint reflected on the augmented map must not be taken into account of localization module. The second map belongs to the map enrichment node and contains all information needed to describe information signs of real environment as virtual footprints on the augmented map. This augmented map is needed for the global path planner node and fetched to it via an appropriate ros topic.