fix: use smart pointers instead of raw pointers
If there are pointers declared in global scope of class and their definitions are initialized inside a local scope of a function, then smart pointer must be used. Otherwise, when the function is executed and completed, a garbage collector removes all local scope variables from a stack memory that causes data lose for the global scope pointers.