feat: create job scheduler service and add job catcher node
The job scheduler service is responsible for receiving requests with initial parameters (sign id, sign coordinates and timestamp when sign was observed). By using them it calculates the rest of parameters, such as the upper time when the sign can be valid. Then the service schedules a job based on the calculated upper time and populates database with an information about the sign and scheduled job. The data stored in the database can be easily obtained using the parameters set in the service response message.

The job catcher node is a listener that intercepts scheduled jobs at the specified time of the job schedule service.

(cherry picked from commit 669057e918838cb49017e7cc988155e69e094931)

# Conflicts:
#	rosbot_description/src/simulation/CMakeLists.txt